+++ packages/terra-form-field/src/terra-dev-site/test/form-field/FieldInline.test.jsx
@@ -1,66 +1,70 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Field from '../../../Field';
+import styles from './FormFieldTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const FieldExamples = () => (
   <div>
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       required
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
 
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       showOptional
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
 
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       required
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
     <br />
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       error="Error Message"
       isInvalid
       required
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
 
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       error="Error Message"
       isInvalid
       showOptional
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
 
     <Field
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isInline
       error="Error Message"
       isInvalid
       required
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
   </div>
 );