+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/FixedMediumSize.jsx
@@ -2,12 +2,16 @@ import React from 'react';
 import Field from 'terra-form-field';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Textarea from 'terra-form-textarea/lib/Textarea';
+import classNames from 'classnames/bind';
+import styles from './FormTextareaDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const FixedMediumSizeExample = () => (
   <Field label="Fixed Textarea - Medium Size" htmlFor="medium">
     <Textarea
       size="medium"
-      style={{ resize: 'none' }}
+      className={cx('textarea')}
       id="medium"
     />
   </Field>