+++ packages/terra-i18n/src/terra-dev-site/doc/i18n/i18n.1.doc.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
+import classNames from 'classnames/bind';
 import ReadMe from '../../../../docs/README.md';
 import { name } from '../../../../package.json';
+import styles from './i18n.1.module.scss';
+
+const cx = classNames.bind(styles);
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -20,8 +24,8 @@ const DocPage = () => (
       {
         title: 'I18nProvider Example',
         description: (
-          <p style={{ fontStyle: 'italic' }}>
-            <span style={{ fontWeight: 'bold' }}>Note:</span>
+          <p className={cx('example-description')}>
+            <span className={cx('weighted-text')}>Note:</span>
             {' '}
             This is an example of a nested I18nProvider. The site overall is an example of using an application-level I18nProvider implemented through the terra-base component.
           </p>