+++ packages/terra-overlay/src/terra-dev-site/doc/example/LoadingOverlayExample.jsx
@@ -4,6 +4,10 @@ import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
 import OverlayContainer from 'terra-overlay/lib/OverlayContainer';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import Button from 'terra-button';
+import classNames from 'classnames/bind';
+import styles from './OverlayDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 class OverlayExample extends React.Component {
   constructor() {
@@ -38,7 +42,7 @@ class OverlayExample extends React.Component {
 
   render() {
     return (
-      <OverlayContainer style={{ height: '100px', width: '100%' }} overlay={this.addLoadingOverlay()}>
+      <OverlayContainer className={cx('overlay-container')} overlay={this.addLoadingOverlay()}>
         <Button text="Trigger Container Overlay" onClick={this.handleTriggerOverlay} />
         <Button text="Trigger FullScreen Overlay" onClick={this.handleTriggerFullScreenOverlay} />
       </OverlayContainer>