+++ packages/terra-status-view/src/terra-dev-site/test/status-view/StatusViewHeightTooSmall.test.jsx
@@ -1,9 +1,13 @@
 import React from 'react';
 import Button from 'terra-button';
+import classNames from 'classnames/bind';
 import StatusView from '../../../StatusView';
+import styles from './StatusViewTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 export default () => (
-  <StatusView id="statusView" variant="error" style={{ border: '1px solid black', height: '250px' }} message="An exception was thrown">
+  <StatusView id="statusView" variant="error" className={cx('status-view-2')} message="An exception was thrown">
     <Button text="OK" key="1" id="button1" />
     <Button text="Cancel" key="2" id="button2" />
   </StatusView>