+++ packages/terra-text/src/terra-dev-site/doc/example/TextWordWrapped.jsx
@@ -1,9 +1,13 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Text from 'terra-text/lib/Text';
+import classNames from 'classnames/bind';
+import styles from './TextDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const TextWordWrapped = () => (
-  <div style={{ width: '150px', border: '1px solid' }}><Text id="text" isWordWrapped>https://engineering.cerner.com/terra-ui/#/components/terra-text/text/text</Text></div>
+  <div className={cx('text-wrapper')}><Text id="text" isWordWrapped>https://engineering.cerner.com/terra-ui/#/components/terra-text/text/text</Text></div>
 );
 
 export default TextWordWrapped;