+++ packages/terra-paginator/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
 2.15.0 - (June 18, 2019)
 ------------------